rough multitasking

+++ Demo/Assets/Graphics/CGincFiles/WindSetup.cginc
@@ -4,7 +4,7 @@ float2 _WindFrequency;
 float _WindStrength;
 float _TileAmount;
 
-fixed2 GetWind(fixed3 center)
+float2 GetWind(float3 center)
 {
 	float2 uv = center.xz * _WindDistortionMap_ST.xy + _WindDistortionMap_ST.zw + _WindFrequency * _Time.y;
 	float2 windSample = tex2Dlod(_WindDistortionMap, float4(uv, 0, 0)).xy; // get wind value ranging (0.33 - 0.66)
@@ -12,6 +12,6 @@ fixed2 GetWind(fixed3 center)
 	float2 windSamplenegpos = windSample01 * 2 - 1; // remap to -1 - 1
 	float2 windSampleStrength = windSamplenegpos * _WindStrength; // multiply by windstrength
 	float2 remap_01 = saturate(windSampleStrength * 0.5 + 0.5); // saturate between 01 to keep low/all/high values dependent on windstrength
-	fixed2 windSampleGrid = floor(remap_01 * (_TileAmount - 0.0001)) * (1 / _TileAmount);
+	float2 windSampleGrid = floor(remap_01 * (_TileAmount - 0.0001)) * (1 / _TileAmount);
 	return windSampleGrid;
 }
