+++ Demo/Assets/Graphics/FlatShade Testing/fragtest.shader
@@ -35,16 +35,16 @@ Shader "Custom/fragtest"
                 v2f o;
                 o.pos = UnityObjectToClipPos (v.vertex);
                 o.worldPos = mul (unity_ObjectToWorld, v.vertex);
-                half3 worldNormal = UnityObjectToWorldNormal (v.normal);
+                float3 worldNormal = UnityObjectToWorldNormal (v.normal);
 
 				o.light = LightCalculation(v.vertex, worldNormal);
 
                 return o;
             }
 			
-			fixed4 frag (v2f i) : SV_Target
+			float4 frag (v2f i) : SV_Target
 			{
-				return fixed4(i.light, i.light, i.light, 1);
+				return float4(i.light, i.light, i.light, 1);
 			}
 			ENDCG
 		}
